requirements.txt
pycountry
requests
bs4
python-dateutil
sqlalchemy
alembic
daterangeparser
dateparser